third_party/blink/web_tests/external/wpt/dom/events/event-global-set-before-handleEvent-lookup.any.worker-expected.txt

This is a testharness.js-based test.
FAIL self.event is set before 'handleEvent' lookup assert_equals: expected (object) object "[object Event]" but got (undefined) undefined
Harness: the test ran to completion.
